+++ src/Configurations/selfconsumptionconfiguration.h
@@ -8,9 +8,9 @@ class SelfConsumptionConfiguration : public QObject
     Q_OBJECT
     Q_PROPERTY(bool selfConsumptionEnabled READ selfConsumptionEnabled WRITE setSelfConsumptionEnabled NOTIFY selfConsumptionEnabledChanged)
     Q_PROPERTY(int selfConsumptionTargetPower READ selfConsumptionTargetPower WRITE setSelfConsumptionTargetPower NOTIFY selfConsumptionTargetPowerChanged)
-    Q_PROPERTY(float selfConsumptionKp READ selfConsumptionKp WRITE setSelfConsumptionKp NOTIFY selfConsumptionKpChanged)
-    Q_PROPERTY(float selfConsumptionKi READ selfConsumptionKi WRITE setSelfConsumptionKi NOTIFY selfConsumptionKiChanged)
-    Q_PROPERTY(float selfConsumptionKd READ selfConsumptionKd WRITE setSelfConsumptionKd NOTIFY selfConsumptionKdChanged)
+    Q_PROPERTY(double selfConsumptionKp READ selfConsumptionKp WRITE setSelfConsumptionKp NOTIFY selfConsumptionKpChanged)
+    Q_PROPERTY(double selfConsumptionKi READ selfConsumptionKi WRITE setSelfConsumptionKi NOTIFY selfConsumptionKiChanged)
+    Q_PROPERTY(double selfConsumptionKd READ selfConsumptionKd WRITE setSelfConsumptionKd NOTIFY selfConsumptionKdChanged)
 
 public:
     explicit SelfConsumptionConfiguration(QObject *parent = nullptr);
@@ -21,28 +21,28 @@ public:
     int selfConsumptionTargetPower() const;
     void setSelfConsumptionTargetPower(int selfConsumptionTargetPower);
 
-    float selfConsumptionKp() const;
-    void setSelfConsumptionKp(float selfConsumptionKp);
+    double selfConsumptionKp() const;
+    void setSelfConsumptionKp(double selfConsumptionKp);
 
-    float selfConsumptionKi() const;
-    void setSelfConsumptionKi(float selfConsumptionKi);
+    double selfConsumptionKi() const;
+    void setSelfConsumptionKi(double selfConsumptionKi);
 
-    float selfConsumptionKd() const;
-    void setSelfConsumptionKd(float selfConsumptionKd);
+    double selfConsumptionKd() const;
+    void setSelfConsumptionKd(double selfConsumptionKd);
 
 signals:
     void selfConsumptionEnabledChanged(bool selfConsumptionEnabled);
     void selfConsumptionTargetPowerChanged(int selfConsumptionTargetPower);
-    void selfConsumptionKpChanged(float selfConsumptionKp);
-    void selfConsumptionKiChanged(float selfConsumptionKi);
-    void selfConsumptionKdChanged(float selfConsumptionKd);
+    void selfConsumptionKpChanged(double selfConsumptionKp);
+    void selfConsumptionKiChanged(double selfConsumptionKi);
+    void selfConsumptionKdChanged(double selfConsumptionKd);
 
 private:
     bool m_selfConsumptionEnabled = false;
     int m_selfConsumptionTargetPower = 0;
-    float m_selfConsumptionKp = 0.10;
-    float m_selfConsumptionKi = 0.10;
-    float m_selfConsumptionKd = 0.00;
+    double m_selfConsumptionKp = 0.10;
+    double m_selfConsumptionKi = 0.10;
+    double m_selfConsumptionKd = 0.00;
 };
 
 #endif // SELFCONSUMPTIONCONFIGURATION_H
